+++ modules/textfile/settings.go
@@ -7,14 +7,17 @@ import (
 
 const defaultTitle = "Textfile"
 
+// Settings defines the configuration properties for this module
 type Settings struct {
 	common *cfg.Common
 
 	filePaths   []interface{}
 	format      bool
 	formatStyle string
+	wrapText    bool
 }
 
+// NewSettingsFromYAML creates a new settings instance from a YAML config block
 func NewSettingsFromYAML(name string, ymlConfig *config.Config, globalConfig *config.Config) *Settings {
 
 	settings := Settings{
@@ -23,6 +26,7 @@ func NewSettingsFromYAML(name string, ymlConfig *config.Config, globalConfig *co
 		filePaths:   ymlConfig.UList("filePaths"),
 		format:      ymlConfig.UBool("format", false),
 		formatStyle: ymlConfig.UString("formatStyle", "vim"),
+		wrapText:    ymlConfig.UBool("wrapText", true),
 	}
 
 	return &settings