+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+### ⚡️ Added
+
+* Textfile module's text wrapping is configurable via the 'wrapText' boolean setting
+
+## 0.11.0
+
+### ⚡️ Added
+
+* GitHub module now supports custom queries for issues and pull requests, by [@Sean-Der](https://github.com/Sean-Der)
+
 ### 🐞 Fixed
 
 * Todoist now properly updates list items when Refresh is called